Added teleport button functionality and fixed small issues

+++ src/main/java/net/laisvall/perspectivemanager/client/util/MathUtil.java
@@ -2,7 +2,7 @@ package net.laisvall.perspectivemanager.client.util;
 
 public class MathUtil {
     public static double roundDouble(double val, int decimals) {
-        float scale = (float) Math.pow(10, decimals);
+        double scale = Math.pow(10, decimals);
         return Math.round(val * scale) / scale;
     }
